Moved tests to unit subfolder
