+++ test/unit/withdrawOptionRequest.ts
@@ -1,8 +1,8 @@
 import { expect } from "chai";
-import { OptionState } from "../src/models/option";
+import { OptionState } from "../../src/models/option";
 
-import { buyer, seller, initializer, dummyOptionRequest, publishDummyOptionRequest } from "./_helpers";
-import { NFTOptContract, deployMainContract } from "../src/utils/deployment";
+import { buyer, seller, initializer, dummyOptionRequest, publishDummyOptionRequest } from "../helpers";
+import { NFTOptContract, deployMainContract } from "../../src/utils/deployment";
 
 describe("withdrawOptionRequest", function () {
     before("prepareEnv", async function () {
